// basic/src/api/types/configurator.ts
export type ConfigurationValue = ValueState & {
  /**
   * The value type.
   */
  type?: 'IntervalValue' | 'SingletonValue';
  excluded?: ValueState;
};
/**
 * Represents a variable in the configuration.
 */
export type ConfigurationVariable = {
  /**
   * The ID of the variable.
   */
  id?: string;
  /**
   * The text to display, representing the variable's name. The text is
   * translated based on the language in the request.
   */
  name?: string;
  /**
   * Type of the `values`.
   */
  valueType?: string;
  /**
   * Number of distinct `values`.
   */
  distinctValueCount?: number;
  /**
   * Indicates whether multiple values can be assigned to the variable
   * simultaneously.
   */
  allowMultipleAssignments?: boolean;
  /**
   * The values that can be assigned to this variable.
   */
  values?: ConfigurationValue[];
  /**
   * Properties that provide additional information about the variable.
   *
   * One way to use properties is to pass information to your application
   * that the variable needs special treatment in the UI. For example, a
   * "Required" property for drawing the user's attention to the variable
   * in the configurator, or a "Show" property indicating whether or not
   * to show the variable in the UI.
   */
  properties?: Property[];
  /**
   * Configurator issues related to the variable. If one of
   * its assigned values is involved in a conflict, `issues` provides
   * information about how to fix the value.
   */
  issues?: Issue[];
};
// ...
export type ConfigureResponse = {
  /**
   * Collection of phases when the underlying product model supports phased configuration.
   */
  phases?: Phase[];
  /**
   * Provides information on completion of the phases.
   */
  phaseCompletion?: {
    /**
     * Indicates the number of phases that have been completed.
     */
    completedPhases?: number;
    /**
     * Indicates whether the phases have been completed (`true`) or if there
     * are still incomplete phases (`false`).
     */
    isComplete?: boolean;
    /**
     * Indicates the total number of phases in the configuration.
     */
    totalPhases?: number;
  };
  /**
   * Collection of sections based on the `viewId` in the request.
   */
  sections?: Section[];
  /**
   * Issues, if any, found while processing the assignments.
   */
  issues?: Issue[];
  priceSheet?: PriceSheet;
  removedAssignments?: RemovedAssignments;
  /**
   * Contains the original request arguments, plus any added by the
   * operation (for example, variants used by the price service).
   */
  arguments?: { [key: string]: any };
  /**
   * Indicates whether all variables in the configuration have an
   * assignment.
   */
  isComplete?: boolean;
  /**
   * Indicates whether the product is configurable (`true`) or is a
   * standard, non-configurable product (`false`).
   */
  isConfigurable?: boolean;
  debug?: Debug;
  /**
   * The language in which `name`s are translated into (for variables,
   * values, sections, and price lines).
   *
   * If requested language was not found, the resolved language is empty.
   * Names are replaced with IDs.
   */
  language?: string;
  /**
   * The `packagePath` from the request, or the latest versioned path of
   * the package if no version was specified. Subsequent API requests
   * during the same configuration session should use the versioned
   * package path.
   */
  packagePath?: string;
};
/**
 * Debug information.
 */
export type Debug = {
  priceTrace?: Trace;
  /**
   * For details on how values are assigned in the configuration.
   */
  solveTrace?: string;
  sublinesTrace?: Trace;
};
export type ErrorResponse = {
  /**
   * The type of error that occurred.
   */
  type?:
    | 'MalformedRequest'
    | 'PackageNotFound'
    | 'CannotLoadPackage'
    | 'MissingProperty'
    | 'NoOrganizationInPackagePath'
    | 'InternalServerError';
  /**
   * A user-friendly description of the error.
   */
  message?: string;
};
/**
 * A variable value that has a lower and upper bounds.
 */
export type IntervalValue = ConfigurationValue & {
  /**
   * Lower bound of the interval.
   */
  lower?: { [key: string]: any };
  /**
   * Upper bound of the interval.
   */
  upper?: { [key: string]: any };
};
/**
 * Represents an issue that occurred during the operation.
 */
export type Issue = {
  /**
   * The issue type.
   */
  type?: string;
  /**
   * An internal message describing the issue.
   */
  message?: string;
  /**
   * Collection of sources that caused the issue. For example, the
   * constraint(s) that could not be met.
   */
  sources?: string[];
  /**
   * Collection of variables affected by the issue, if any. Assigning
   * different values to the variables may help fix the issue.
   */
  affectedVariables?: string[];
};
// ...
export type PriceLine = {
  /**
   * The ID of the price line. Used in a price line assignment to
   * reference a price line.
   */
  id?: string;
  /**
   * The ID of the price procedure step that generated the price line.
   */
  stepId?: string;
  /**
   * A name describing the price line. The text is translated based on
   * the language in the request.
   */
  name?: string;
  /**
   * Optionally contains the feature that caused this price line to be
   * added.
   */
  feature?: string;
  rate?: PriceLineRate;
  result?: ValueWithUnit;
  status?: PriceLineStatus;
  /**
   * Additional properties for this price line, if any.
   */
  properties?: Property[];
};
// ...
export type PriceSheet = {
  /**
   * The internal ID.
   */
  id?: string;
  totals?: PriceSummary;
  /**
   * The price lines comprising the sheet. When `onlyPriceTotals` is
   * `true`, price lines are not provided.
   */
  priceLines?: PriceLine[];
  /**
   * Indicates whether any of the price lines are invalid. This can
   * happen, for example, due to a mismatch of units, or a division by
   * zero error.
   */
  isInvalid?: boolean;
};
/**
 * The summary of a `PriceSheet`, including the product's net value, tax, and total price (net value minus tax).
 */
export type PriceSummary = {
  tax?: ValueWithCurrency;
  net?: ValueWithCurrency;
  total?: ValueWithCurrency;
};
// ...
export type Property = {
  /**
   * ID of the property.
   */
  id?: string;
  /**
   * > For single-valued properties only.
   *
   * The property value. For example, 'true', 42, or 'This is a
   * description'.
   */
  value?: { [key: string]: any };
  /**
   * > For properties with multiple values only.
   *
   * The property values. For example {'France', 'Denmark'}.
   */
  values?: { [key: string]: any }[];
  /**
   * The property type, which can be:
   *
   * - `String`: For example, `LongDescription="This a description"`.
   * - `Number`: For example, `Weight=42.0`.
   * - `Date`: Supports the ISO 8601 date time format. For example,
   *   `Expires=2020-05-03`.
   * - `Uri`: For example, `Image="http://placehold.it/350x150"`.
   * - `Boolean`: For example, `Show=false`.
   */
  type?: 'String' | 'Number' | 'Date' | 'Uri' | 'Boolean';
};
/**
 * Variable and/or price line assignments that were removed from a
 * configure and/or price request.
 */
export type RemovedAssignments = {
  /**
   * A collection of assignments removed during the configuration solve.
   *
   * An assignment is removed when:
   *
   * - Other assignments are made with values that are incompatible.
   * - Multiple assignments are made to the same variable, and the
   *   variable does not allow multiple assignments
   *   (`allowMultipleAssignments`) is `false`). In this case, the last
   *   assignment is preserved, and all others are removed.
   *
   * For SAP products, assignments with values that are `incompatible`
   * are not added to `removedAssignments`.
   */
  variableAssignments?: RemovedAssignment[];
  /**
   * A collection of assignments removed by the price service.
   *
   * Assignments are removed when they have been made to non-existing
   * price lines. This can happen to assignments to price lines of
   * variant that was removed during configuration.
   *
   * Clients should use these values to update their state.
   */
  priceLineAssignments?: PriceLineAssignment[];
};
/**
 * Represents a removed value assignment to a variable.
 */
export type RemovedAssignment = {
  variable?: RemovedVariable;
  value?: RemovedValue;
};
// ...
export type Section = {
  /**
   * The ID of the section.
   */
  id?: string;
  /**
   * The text to display, representing the section's name. The text is
   * translated based on the language in the request.
   */
  name?: string;
  /**
   * The section's subsections, if any.
   */
  sections?: Section[];
  /**
   * The collection of ConfigurationVariable objects belonging to this
   * section.
   */
  variables?: ConfigurationVariable[];
  /**
   * Miscellaneous properties.
   */
  properties?: Property[];
};
export type Phase = {
  /**
   * Indicates whether the phase has been completed.
   */
  isComplete?: boolean;
  /**
   * The sections of the phase.
   */
  sections?: Section[];
};
// ...
export type SingletonValue = ConfigurationValue & {
  /**
   * A textual representation of the value. For non-numeric values,
   * the text is translated based on the language in the request.
   */
  name?: string;
  /**
   * The value, which may be a string, number, or date. For example,
   * "RED", 42, 2020-05-03.
   */
  value?: string | number;
  /**
   * Properties providing additional information about the value.
   */
  properties?: Property[];
};
// ...
export type Trace = {
  identifier?: string;
  message?: string;
  details?: string;
  severity?: string;
  children?: Trace[];
};
// ...
export type ValueState = {
  /**
   * Describes how the value was assigned. If this property is not
   * include in the response, the value is unassigned.
   *
   * | Value | Description |
   * |-------|-------------|
   * |`byUser`|The user assigned the value.|
   * |`byDefault`|The system assigned a default value.|
   * |`byPhase`|The system assigned a value due to a previous phase assignment.|
   * |`byRule`|The system assigned the value due to a rule.|
   */
  assigned?: 'byUser' | 'byDefault' | 'byPhase' | 'byRule';
  /**
   * Indicates whether the value is incompatible. Assigning incompatible
   * values causes other assignments to be removed.
   */
  incompatible?: boolean;
  /**
   * The configuration state of a variable's value. Describes whether the
   * value is assigned, can be assigned, and the consequences of
   * assigning it.
   *
   * `state` is set only if `Settings.includeStateAndJustification` is
   * `true`. Combined with `Justification` it details
   * `ValueState.assigned` and `ValueState.incompatible`.
   *
   * | Assignment state | Description |
   * |------------------|-------------|
   * | Available | The value can be freely assigned. |
   * | Unavailable | Depending on the `justification`, the value can either not be assigned at all, or would remove one of the other assigned values.|
   * | Selected | The value is assigned, either by the user, or because it is a default value.|
   * | Inferred | The value is assigned for one of these reasons: As a consequence of an assignment or a default applied to another variable; It is the only value available according to the rules; It is the only value defined in the variable's domain.|
   */
  state?: 'Available' | 'Unavailable' | 'Selected' | 'Inferred';
  /**
   * The justification of a value's state. Only set if
   * `includeStateAndJustification` is `true`.
   *
   * Provides context for how the configuration engine determined the
   * value's state. Combined with `State` it details
   * `ValueState.assigned` and `ValueState.incompatible`.
   *
   * | Justification | Description |
   * |------------------|-------------|
   * | None | A value with state Available has no justification. |
   * | Rule | The state is exclusively due to rules, and not because of other assignments. |
   * | Assignment | If state is Selected, it is because the value was assigned in the request. If state is Unavailable or Inferred, it is due to an assignment to other variable(s) in combination with the rules. |
   * | Default | If state is Selected, it is because it is a default value. If state is Unavailable or Inferred, it is because another default value was assigned. |
   * | Phase | The value state is due to assignments, defaults, and/or rules from a previous phase. Note: Only applies to state Unavailable or Inferred, only when the phases are defined in the model or request. |
   */
  justification?: 'None' | 'Rule' | 'Assignment' | 'Default' | 'Phase';
};
/**
 * A value and its currency.
 */
export type ValueWithCurrency = {
  /**
   * The value.
   */
  value?: number;
  /**
   * The currency. For example, "EU" for euros, or "USD" for US dollars.
   */
  currency?: string;
};
/**
 * A value and its unit of measurement. When representing batch sizes,
 * `value` is the size, and `unit` is the batch.
 */
export type ValueWithUnit = {
  /**
   * The value.
   */
  value?: number;
  /**
   * The unit of measurement.
   */
  unit?: string;
};